+++ application/logs/CHANGELOG.md
@@ -8,6 +8,7 @@
 - bugfix: a kapcsolódó tartalmak cím ne jelenjen meg, ha van meta, de nincs szerző vagy sorozat
 - bugfix: hiányzó favicon pótlása
 - bugfix: hiányzó lapozás a cikkek szerkesztése lista alján
+- "Vissza a cikkek listájához" gomb a cikkek szerkesztéséhez
 
 ## RELEASE 2018.11.24.